comments tree padding revised

+++ zavtra/corecontent/templatetags/utils.py
@@ -4,7 +4,7 @@ register = Library()
 
 @register.filter
 def mul(val, k):
-    return int(val*float(k))
+    return val*float(k)
 
 @register.filter
 def negate(val):
